Admin Side course fetch with module and lessons

+++ src/main/java/com/example/demo/repository/CourseModuleRepository.java
@@ -10,6 +10,7 @@ public interface CourseModuleRepository extends JpaRepository<CourseModule, Inte
 	
 	List<CourseModule> findByCourseCourseIdOrderByPosition(Integer courseId);
 
-	int countByCourseCourseId(Integer courseId);
+	Long countByCourseCourseId(Long courseId);
+	
 
 }
